# schema/openapi.yaml
openapi: 3.1.0
info:
  title: Energetic Paradigm API
  version: 0.1.0
  summary: Provider-independent EP API boundary
  description: |
    Formal external contract for the current EP API boundary.
    This file defines the stable request and response shapes that callers use.
    Provider-specific payloads stay inside connectors and are not exposed here.

servers:
  - url: http://127.0.0.1:8010

tags:
  - name: ep
    description: EP request/response endpoints
  - name: system
    description: Health and provider status endpoints

paths:
  /v1/ep/respond:
    post:
      tags: [ep]
      operationId: epRespond
      summary: Produce a final EP answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EPRequestEnvelope'
            examples:
              unified:
                summary: Full unified request
                value:
                  request_id: req_20260312_0001
                  client:
                    client_id: openclaw-demo
                    client_type: partner_app
                    session_id: sess_abc123
                  user:
                    user_id: user_001
                    locale: zh-Hant
                    timezone: Asia/Riyadh
                  input:
                    text: Please analyze this problem with Energetic Paradigm.
                    attachments: []
                  ep:
                    mode: auto
                    depth: auto
                    domain: general
                    format: structured
                    safety_level: standard
                  routing:
                    preferred_provider: auto
                    allowed_providers: [openai, local]
                    fallback_provider: local
                    allow_fallback: true
                  output:
                    response_type: final_answer
                    max_tokens: 1200
                    temperature: 0.2
                    include_citations: false
                    include_trace: false
                    language: zh-Hant
                  metadata:
                    source: terminal
                    tags: [ep, api-test]
              minimal:
                summary: Minimal request
                value:
                  input:
                    text: Please analyze this problem.
                  ep:
                    mode: auto
                    depth: auto
                  routing:
                    preferred_provider: auto
                    allowed_providers: [openai, local]
                  output:
                    language: zh-Hant
      responses:
        '200':
          description: Unified EP response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EPResponseEnvelope'
              examples:
                success:
                  summary: Success
                  value:
                    request_id: req_20260312_0001
                    status: success
                    result:
                      response_type: final_answer
                      content: This is the formal answer after EP routing and provider orchestration.
                      language: zh-Hant
                      format: structured
                    ep:
                      mode_applied: deep
                      depth_applied: long
                      domain_applied: general
                      routing_path: ep_deep_reasoning
                    provider:
                      selected: openai
                      model: gpt-5
                      fallback_used: false
                      simulated: false
                    usage:
                      input_tokens: 850
                      output_tokens: 1200
                      total_tokens: 2050
                      estimated_cost: 0.0
                    timing:
                      started_at: 2026-03-12T18:10:00+03:00
                      completed_at: 2026-03-12T18:10:06+03:00
                      latency_ms: 6120
                    warnings: []
                    error: null
                error:
                  summary: Error
                  value:
                    request_id: req_20260312_0002
                    status: error
                    result: null
                    ep:
                      mode_applied: auto
                      depth_applied: auto
                      domain_applied: general
                      routing_path: null
                    provider:
                      selected: local
                      model: null
                      fallback_used: true
                      simulated: false
                    usage:
                      input_tokens: 0
                      output_tokens: 0
                      total_tokens: 0
                      estimated_cost: 0.0
                    timing:
                      started_at: 2026-03-12T18:12:00+03:00
                      completed_at: 2026-03-12T18:12:02+03:00
                      latency_ms: 2100
                    warnings:
                      - primary_provider_failed:openai
                    error:
                      code: PROVIDER_UNAVAILABLE
                      message: Selected provider is temporarily unavailable.
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyEnvelope'
              example:
                status: error
                error:
                  code: INVALID_REQUEST
                  message: input.text is required

  /v1/ep/analyze:
    post:
      tags: [ep]
      operationId: epAnalyze
      summary: Produce an EP analysis response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EPRequestEnvelope'
      responses:
        '200':
          description: Unified EP analysis response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EPResponseEnvelope'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyEnvelope'

  /v1/ep/health:
    get:
      tags: [system]
      operationId: epHealth
      summary: Get service health and provider readiness
      responses:
        '200':
          description: Health information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/ep/providers:
    get:
      tags: [system]
      operationId: epProviders
      summary: Get provider status list
      responses:
        '200':
          description: Provider list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvidersResponse'

components:
  schemas:
    EPRequestEnvelope:
      oneOf:
        - $ref: '#/components/schemas/EPUnifiedRequest'
        - $ref: '#/components/schemas/MinimalRequest'
      description: Unified external request boundary. Callers may send either the full unified request or the minimal request.

    EPUnifiedRequest:
      type: object
      additionalProperties: false
      required:
        - input
      properties:
        request_id:
          type: string
          description: Unique request identifier for tracing and logs.
        client:
          $ref: '#/components/schemas/ClientInfo'
        user:
          $ref: '#/components/schemas/UserInfo'
        input:
          $ref: '#/components/schemas/InputBlock'
        ep:
          $ref: '#/components/schemas/EPBlock'
        routing:
          $ref: '#/components/schemas/RoutingBlock'
        output:
          $ref: '#/components/schemas/OutputBlock'
        metadata:
          $ref: '#/components/schemas/MetadataBlock'

    MinimalRequest:
      type: object
      additionalProperties: false
      required:
        - input
      properties:
        input:
          $ref: '#/components/schemas/MinimalInput'
        ep:
          $ref: '#/components/schemas/MinimalEP'
        routing:
          $ref: '#/components/schemas/MinimalRouting'
        output:
          $ref: '#/components/schemas/MinimalOutput'

    ClientInfo:
      type: object
      additionalProperties: false
      properties:
        client_id:
          type: string
          default: default-client
        client_type:
          $ref: '#/components/schemas/ClientType'
        session_id:
          type: string
          default: session-default
      default:
        client_id: default-client
        client_type: terminal
        session_id: session-default

    UserInfo:
      type: object
      additionalProperties: false
      properties:
        user_id:
          type: string
          default: anonymous
        locale:
          type: string
          default: zh-Hant
        timezone:
          type: string
          default: Asia/Riyadh
      default:
        user_id: anonymous
        locale: zh-Hant
        timezone: Asia/Riyadh

    InputBlock:
      type: object
      additionalProperties: false
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
        attachments:
          type: array
          items:
            type: object
            additionalProperties: true
          default: []

    EPBlock:
      type: object
      additionalProperties: false
      properties:
        mode:
          $ref: '#/components/schemas/EPMode'
        depth:
          $ref: '#/components/schemas/EPDepth'
        domain:
          type: string
          default: general
        format:
          $ref: '#/components/schemas/EPFormat'
        safety_level:
          $ref: '#/components/schemas/SafetyLevel'
      default:
        mode: auto
        depth: auto
        domain: general
        format: structured
        safety_level: standard

    RoutingBlock:
      type: object
      additionalProperties: false
      properties:
        preferred_provider:
          $ref: '#/components/schemas/ProviderNameOrAuto'
        allowed_providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderName'
          default: [openai, local]
        fallback_provider:
          $ref: '#/components/schemas/ProviderName'
        allow_fallback:
          type: boolean
          default: true
      default:
        preferred_provider: auto
        allowed_providers: [openai, local]
        fallback_provider: local
        allow_fallback: true

    OutputBlock:
      type: object
      additionalProperties: false
      properties:
        response_type:
          $ref: '#/components/schemas/ResponseType'
        max_tokens:
          type: integer
          minimum: 1
          default: 1200
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.2
        include_citations:
          type: boolean
          default: false
        include_trace:
          type: boolean
          default: false
        language:
          type: string
          default: zh-Hant
      default:
        response_type: final_answer
        max_tokens: 1200
        temperature: 0.2
        include_citations: false
        include_trace: false
        language: zh-Hant

    MetadataBlock:
      type: object
      additionalProperties: false
      properties:
        source:
          type: string
          default: terminal
        tags:
          type: array
          items:
            type: string
          default: []
      default:
        source: terminal
        tags: []

    MinimalInput:
      type: object
      additionalProperties: false
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1

    MinimalEP:
      type: object
      additionalProperties: false
      properties:
        mode:
          $ref: '#/components/schemas/EPMode'
        depth:
          $ref: '#/components/schemas/EPDepth'
      default:
        mode: auto
        depth: auto

    MinimalRouting:
      type: object
      additionalProperties: false
      properties:
        preferred_provider:
          $ref: '#/components/schemas/ProviderNameOrAuto'
        allowed_providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderName'
          default: [openai, local]
      default:
        preferred_provider: auto
        allowed_providers: [openai, local]

    MinimalOutput:
      type: object
      additionalProperties: false
      properties:
        language:
          type: string
          default: zh-Hant
      default:
        language: zh-Hant

    EPResponseEnvelope:
      type: object
      additionalProperties: false
      required:
        - request_id
        - status
        - result
        - ep
        - provider
        - usage
        - timing
        - warnings
        - error
      properties:
        request_id:
          type: string
        status:
          $ref: '#/components/schemas/Status'
        result:
          oneOf:
            - $ref: '#/components/schemas/ResultBlock'
            - type: 'null'
        ep:
          $ref: '#/components/schemas/AppliedEPBlock'
        provider:
          $ref: '#/components/schemas/ProviderBlock'
        usage:
          $ref: '#/components/schemas/UsageBlock'
        timing:
          $ref: '#/components/schemas/TimingBlock'
        warnings:
          type: array
          items:
            type: string
        error:
          oneOf:
            - $ref: '#/components/schemas/ErrorBlock'
            - type: 'null'
        trace:
          oneOf:
            - $ref: '#/components/schemas/TraceBlock'
            - type: 'null'

    ResultBlock:
      type: object
      additionalProperties: false
      required:
        - response_type
        - content
        - language
        - format
      properties:
        response_type:
          $ref: '#/components/schemas/ResponseType'
        content:
          type: string
        language:
          type: string
        format:
          type: string

    AppliedEPBlock:
      type: object
      additionalProperties: false
      required:
        - mode_applied
        - depth_applied
        - domain_applied
        - routing_path
      properties:
        mode_applied:
          oneOf:
            - $ref: '#/components/schemas/EPMode'
            - type: string
        depth_applied:
          oneOf:
            - $ref: '#/components/schemas/EPDepth'
            - type: string
        domain_applied:
          type: string
        routing_path:
          oneOf:
            - type: string
            - type: 'null'

    ProviderBlock:
      type: object
      additionalProperties: false
      required:
        - selected
        - model
        - fallback_used
      properties:
        selected:
          oneOf:
            - $ref: '#/components/schemas/ProviderName'
            - type: string
        model:
          oneOf:
            - type: string
            - type: 'null'
        fallback_used:
          type: boolean
        simulated:
          type: boolean
          default: false

    UsageBlock:
      type: object
      additionalProperties: false
      required:
        - input_tokens
        - output_tokens
        - total_tokens
        - estimated_cost
      properties:
        input_tokens:
          type: integer
          minimum: 0
        output_tokens:
          type: integer
          minimum: 0
        total_tokens:
          type: integer
          minimum: 0
        estimated_cost:
          type: number
          minimum: 0

    TimingBlock:
      type: object
      additionalProperties: false
      required:
        - started_at
        - completed_at
        - latency_ms
      properties:
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        latency_ms:
          type: integer
          minimum: 0

    ErrorBlock:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string

    TraceBlock:
      type: object
      additionalProperties: false
      properties:
        client_id:
          type: string
        provider_requested:
          oneOf:
            - $ref: '#/components/schemas/ProviderNameOrAuto'
            - type: string
        provider_allowed:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/ProviderName'
              - type: string
        fallback_provider:
          oneOf:
            - $ref: '#/components/schemas/ProviderName'
            - type: string

    ErrorOnlyEnvelope:
      type: object
      additionalProperties: false
      required:
        - status
        - error
      properties:
        status:
          type: string
          enum: [error]
        error:
          $ref: '#/components/schemas/ErrorBlock'

    HealthResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - service
        - version
        - openai_configured
        - providers
      properties:
        status:
          type: string
          enum: [ok]
        service:
          type: string
        version:
          type: string
        openai_configured:
          type: boolean
        providers:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ProviderRuntimeStatus'

    ProvidersResponse:
      type: object
      additionalProperties: false
      required:
        - providers
      properties:
        providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderListItem'

    ProviderListItem:
      type: object
      additionalProperties: false
      required:
        - name
        - status
      properties:
        name:
          oneOf:
            - $ref: '#/components/schemas/ProviderName'
            - type: string
        status:
          $ref: '#/components/schemas/ProviderRuntimeStatus'

    ProviderRuntimeStatus:
      type: string
      enum:
        - ready
        - not_configured
        - declared_only

    ClientType:
      type: string
      enum:
        - web_app
        - partner_app
        - internal_tool
        - terminal

    EPMode:
      type: string
      enum:
        - auto
        - standard
        - deep
        - domain

    EPDepth:
      type: string
      enum:
        - auto
        - short
        - medium
        - long

    EPFormat:
      type: string
      enum:
        - plain
        - structured
        - executive
        - technical

    SafetyLevel:
      type: string
      enum:
        - standard
        - strict

    ResponseType:
      type: string
      enum:
        - final_answer
        - analysis
        - summary

    Status:
      type: string
      enum:
        - success
        - error

    ProviderNameOrAuto:
      type: string
      enum:
        - auto
        - openai
        - gemini
        - deepseek
        - openclaw
        - kimly
        - local

    ProviderName:
      type: string
      enum:
        - openai
        - gemini
        - deepseek
        - openclaw
        - kimly
        - local

    ErrorCode:
      type: string
      enum:
        - INVALID_REQUEST
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - RATE_LIMITED
        - INTERNAL_ERROR
        - PROVIDER_ERROR
        - PROVIDER_UNAVAILABLE
        - PROVIDER_TIMEOUT